First Commit:
Displayes a card facing back, tap on it to draw the next random card from the deck. Tap on it again to face the back of the card. Tap again to draw randome card from deck again. This goes on until the entire deck has been drawn. After that displays a blank deck.